fix(ci): split internal/watch/dedup_test.go under the 800-line hygiene gate
The two new currency-corruption regression tests pushed this file to 846
lines. Split along the file's own theme boundary: dedup_test.go keeps
Store.Add dedup/currency/history-eviction/route-TTL tests; the new
store_lifecycle_test.go gets scheduler-lock, Migrate, and Load tests.
Pure move (plus each file's own now-correct import list), no logic
changes. Same pattern already used earlier in this PR for
watch.go/store.go.

Verified: go build/vet/staticcheck/golangci-lint clean, scripts/ci/
check-file-size.sh passes locally (1553 files checked, only the
allowlisted CHANGELOG.md over 800 lines), full `go test ./...` green.
